requirements.txt

fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.8.2

# GraphQL
strawberry-graphql[fastapi]==0.243.0

# Database
sqlalchemy[asyncio]==2.0.35
psycopg[binary,pool]==3.2.3

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0